Set port speed under interfaces
This patch covers a fix on juniper connected interfaces
to explicitly set the juniper connected port speed to the
speed defined in the quads database for the interface(s)
of a host.

When defining an interface, if speed is not passed as an
argument, then the interface speed is set to 0 in the database.

If the interface speed is set to 0, then there
will be no additional setting.  Otherwise, speed is looked
up and set accordingly. For example, if the speed is set
to 25 in the associated quads data for the interface, then
the switch will get the additional:

   set interfaces <interface name> speed 25g

Fixes: https://github.com/redhat-performance/quads/issues/566
Change-Id: I7c2d3788bb495eea6d08dbd43c4451e01c570823